+++ backend/src/GatorGuide/database/__init__.py
@@ -1,0 +1,2 @@
+from .gather_courses import populate_courses
+from .parse_prereqs import parse
