+++ backend/src/GatorGuide/database/gather_courses.py
@@ -3,22 +3,19 @@ from pydantic.json import pydantic_encoder
 from GatorGuide.database.models import Course
 
 from sqlmodel import Field, Relationship, Session, SQLModel, create_engine, select
+import sqlmodel
 import GatorGuide.database.models as models
 import pathlib
-from GatorGuide.database.parse_prereqs import parse
-
-sqlite_file_name = pathlib.Path(__file__).parent.resolve().joinpath("./database.db")
-sqlite_url = f"sqlite:///{sqlite_file_name}"
-
-engine = create_engine(sqlite_url)
+from GatorGuide.database import parse
 
 
-last_num = 0  # used by the one.uf api to get the next page of courses
+def populate_courses(session: Session):
+    last_num = 0  # used by the one.uf api to get the next page of courses
 
-courses: list[Course] = []
-course_json = []
-count = 0  # used to keep track of how many courses have been registered
-with Session(engine) as session:
+    courses: list[Course] = []
+    course_json = []
+    count = 0  # used to keep track of how many courses have been registered
+
     while True:
         URL = f"https://one.uf.edu/apix/soc/schedule?ai=false&auf=false&category=CWSP&class-num=&course-code=&course-title=&cred-srch=&credits=&day-f=&day-m=&day-r=&day-s=&day-t=&day-w=&dept=&eep=&fitsSchedule=false&ge=&ge-b=&ge-c=&ge-d=&ge-h=&ge-m=&ge-n=&ge-p=&ge-s=&instructor=&last-control-number={last_num}&level-max=&level-min=&no-open-seats=false&online-a=&online-c=&online-h=&online-p=&period-b=&period-e=&prog-level=&qst-1=&qst-2=&qst-3=&quest=false&term=2251&wr-2000=&wr-4000=&wr-6000=&writing=false&var-cred=&hons=false"
 
@@ -65,7 +62,6 @@ with Session(engine) as session:
                     )
                     if course is not None:
                         prereqs[-1].courses.append(course)
-        if(c["code"] == "COP3503")
         session.add(
             Course(
                 code=c["code"],
@@ -77,6 +73,13 @@ with Session(engine) as session:
         )
 
     session.commit()
-# with open("courses.json", "w") as f:
-#     json.dump(courses, f, default=pydantic_encoder)
-#     print(f"TOTAL COUNT: {count}")
+
+
+if __name__ == "__main__":
+    sqlite_file_name = pathlib.Path(__file__).parent.resolve().joinpath("./database.db")
+    sqlite_url = f"sqlite:///{sqlite_file_name}"
+
+    engine = create_engine(sqlite_url)
+
+    with Session(engine) as session:
+        populate_courses(session)
