+++ backend/src/GatorGuide/database/operations.py
@@ -1,0 +1,1 @@
+# ^(?!.*(COP3520)).*^(COP3|MHF4) useful regex
